split up seer tests
